Refactor test according to new specs
